Move selector into its own package
