+++ mongodb/src/test/scala/quasar/physical/mongodb/optimize.scala
@@ -22,6 +22,7 @@ import quasar.common.SortDir
 import quasar.physical.mongodb.accumulator._
 import quasar.physical.mongodb.expression._
 import quasar.physical.mongodb.optimize.pipeline._
+import quasar.physical.mongodb.selector.Selector
 import quasar.physical.mongodb.workflow._
 
 import scalaz._, Scalaz._
