+++ mongodb/src/test/scala/quasar/physical/mongodb/pipeline.scala
@@ -21,6 +21,7 @@ import slamdata.Predef._
 import quasar.common.SortDir
 import quasar.physical.mongodb.accumulator._
 import quasar.physical.mongodb.expression._
+import quasar.physical.mongodb.selector.Selector
 import quasar.physical.mongodb.workflow._
 
 import org.scalacheck._
